Move notification from pkg to model
