+++ pkg/oauth/client.go
@@ -10,12 +10,12 @@ import (
 
 	"github.com/cozy/cozy-stack/pkg/registry"
 
+	"github.com/cozy/cozy-stack/model/notification"
 	"github.com/cozy/cozy-stack/pkg/consts"
 	"github.com/cozy/cozy-stack/pkg/couchdb"
 	"github.com/cozy/cozy-stack/pkg/couchdb/mango"
 	"github.com/cozy/cozy-stack/pkg/crypto"
 	"github.com/cozy/cozy-stack/pkg/instance"
-	"github.com/cozy/cozy-stack/pkg/notification"
 	"github.com/cozy/cozy-stack/pkg/permissions"
 
 	jwt "gopkg.in/dgrijalva/jwt-go.v3"
